Basic Mechanics
Added the ability to throw a character to a location and have the force
of gravity act upon him. Can now switch between characters and move them
when appropriate.

Bugs :-
Nigel, after being thrown will bounce on the screen
Switching characters will sometimes not work as intended if the user is
hovering over the buttons