Tile-based levels
Per Pixel collision was not working correctly, game now loads in a
tile-based map.
As a result, the frame rate has slowed down. The solution to this would
be to create a texture manager and if i tile ahs already been loaded in,
return that pre-existing texture.

Collisions can now be triggered on walls (nothing stops you from going
through them however).